backend/requirements.txt
openai
boto3
requests
python-dotenv
openai"[embeddings]"
pyarrow
pinecone